feat: adding zip as an output
Zip can now be selected as an output format, the compiled binary will be zipped up which can be directly uploaded to create an AWS Lambda function.

The `--output-format` paramters controls the output format, the two
current options are `Zip` and `Binary` with `Binary` being the default.

Example usage to create a bootstrap zip.

```
cargo lambda build --output-format Zip
```