add support for extensions
this change allows the CLI to generate a stub package for either an extension or a
logs extension in addition to the default function.  the caller specifies
`--type extension` or `--type logs_extension` if desired.  with this change we
do some validation against other arguments; for instance, it makes no sense for
the caller to specify `--http` for an extension.  we fail with a helpful error
message in this case.  this change is accompanied by a corresponding change to
the default template repo.