Build lambda respecting CARGO_TARGET_DIR
If no `lambda-dir` parameter is given but a non-default `target-dir` is
configured either in the project config or as environment variable,
`cargo-lambda` so far failed to build the `bootstrap.zip` because it
assumed the target dir to be always `./target`. This commit adds a
helper function to extract the target directory from cargo's metadata,
resolving the config key and environment variable in the right order.