discover ignore files from parent origins
this allows us to ignore files that are declared in a `.gitignore` file in a
parent directory.  an example of this is in the [aws-lambda-rust-runtime](https://github.com/awslabs/aws-lambda-rust-runtime)
crate, where the root crate's `.gitignore` file ignores all `target/` directories
so that crates in the `examples/` directory need not ignore the target themselves.