feat(new): Custom templates with configuration prompts.
Implement new configuration for templates that allow users to add
their own prompts when the template is used.

Allow to put the template files into a subdirectory to create more involved templates that separate the template files themselves and other files used to explain what the template is for and how to use it.