feat(watch): TLS connections (#727)
* feat(watch): TLS server implementation.

Allow the watch server to configure a TLS connection when
the tls flags are provided. This will make the server to listen to HTTPS connections behind the same port.

* feat(invoke): Allow to call local functions behind a TLS connection.

* feat(watch): Use a TLS proxy in front of the runtime server.

The lambda runtime embedded in each function expects the runtime server to listen on localhost.
Instead of binding the runtime server to TLS connections,
we put start a proxy TLS server that handles the secure connection
and forwards requests to the runtime server.

* docs(watch): Add documentation for the TLS support.

* chore(watch): Format imports.

* feat(watch): Allow to configure global TLS configuration.

* fix(watch): Call install_provider only once.